// src/renderer/shared/api/permission/common/types.ts
export enum OperationType {
  TRANSFER = 'transfer',
  RECEIVE = 'receive',
  STAKING = 'staking',
  CREATE_MULTISIG_TX = 'createMultisigTx',
  APPROVE_MULTISIG_TX = 'approveMultisigTx',
  REJECT_MULTISIG_TX = 'rejectMultisigTx',
  CREATE_ANY_PROXY = 'createAnyProxy',
  CREATE_NON_ANY_PROXY = 'createNonAnyProxy',
  REMOV_PROXY = 'removeProxy',
}